Fix Chakra UI dependency bundling issue with Vite.
