--- playground/chakra-ui-playground/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "chakra-ui-playground",
-  "version": "1.0.0",
-  "description": "A comprehensive Chakra UI playground for RedwoodSDK",
-  "main": "index.js",
-  "type": "module",
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "build": "vite build",
-    "dev": "vite dev",
-    "dev:init": "rw-scripts dev-init",
-    "preview": "vite preview",
-    "worker:run": "rw-scripts worker-run",
-    "clean": "npm run clean:vite",
-    "clean:vite": "rm -rf ./node_modules/.vite",
-    "release": "rw-scripts ensure-deploy-env && npm run clean && npm run build && RWSDK_DEPLOY=1 wrangler deploy",
-    "generate": "rw-scripts ensure-env && wrangler types",
-    "check": "npm run generate && npm run types",
-    "types": "tsc",
-    "test": "vitest run --config ../vitest.config.mts"
-  },
-  "dependencies": {
-    "rwsdk": "workspace:*",
-    "react": "19.2.0-canary-3fb190f7-20250908",
-    "react-dom": "19.2.0-canary-3fb190f7-20250908",
-    "react-server-dom-webpack": "19.2.0-canary-3fb190f7-20250908",
-    "@chakra-ui/react": "^3.2.0",
-    "@emotion/react": "^11.13.3",
-    "@emotion/styled": "^11.13.0"
-  },
-  "devDependencies": {
-    "@cloudflare/vite-plugin": "1.12.4",
-    "@cloudflare/workers-types": "4.20250913.0",
-    "@types/node": "22.14.0",
-    "@types/react": "19.1.2",
-    "@types/react-dom": "19.1.2",
-    "typescript": "5.8.3",
-    "vite": "7.1.5",
-    "wrangler": "4.35.0"
-  },
-  "pnpm": {
-    "onlyBuiltDependencies": [
-      "esbuild",
-      "sharp",
-      "workerd"
-    ]
-  }
-}
