--- playground/chakra-ui-playground/src/app/Document.tsx
@@ -1,30 +1,0 @@
-import stylesUrl from "./styles.css?url";
-export const Document: React.FC<{ children: React.ReactNode }> = ({
-  children,
-}) => (
-  <html lang="en">
-    <head>
-      <meta charSet="utf-8" />
-      <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <title>Chakra UI Playground - RedwoodSDK</title>
-      <link rel="modulepreload" href="/src/client.tsx" />
-      <link rel="stylesheet" href={stylesUrl} />
-      <link
-        rel="icon"
-        type="image/svg+xml"
-        href="/favicon-dark.svg"
-        media="(prefers-color-scheme: dark)"
-      />
-      <link
-        rel="icon"
-        type="image/svg+xml"
-        href="/favicon-light.svg"
-        media="(prefers-color-scheme: light)"
-      />
-    </head>
-    <body>
-      <div id="root">{children}</div>
-      <script>import("/src/client.tsx")</script>
-    </body>
-  </html>
-);
