--- playground/chakra-ui-playground/src/app/headers.ts
@@ -1,31 +1,0 @@
-import { RouteMiddleware } from "rwsdk/router";
-
-export const setCommonHeaders =
-  (): RouteMiddleware =>
-  ({ response, rw: { nonce } }) => {
-    if (!import.meta.env.VITE_IS_DEV_SERVER) {
-      // Forces browsers to always use HTTPS for a specified time period (2 years)
-      response.headers.set(
-        "Strict-Transport-Security",
-        "max-age=63072000; includeSubDomains; preload",
-      );
-    }
-
-    // Forces browser to use the declared content-type instead of trying to guess/sniff it
-    response.headers.set("X-Content-Type-Options", "nosniff");
-
-    // Stops browsers from sending the referring webpage URL in HTTP headers
-    response.headers.set("Referrer-Policy", "no-referrer");
-
-    // Explicitly disables access to specific browser features/APIs
-    response.headers.set(
-      "Permissions-Policy",
-      "geolocation=(), microphone=(), camera=()",
-    );
-
-    // Defines trusted sources for content loading and script execution:
-    response.headers.set(
-      "Content-Security-Policy",
-      `default-src 'self'; script-src 'self' 'nonce-${nonce}' https://challenges.cloudflare.com; style-src 'self' 'unsafe-inline'; frame-ancestors 'self'; frame-src 'self' https://challenges.cloudflare.com https://rwsdk.com; object-src 'none';`,
-    );
-  };
