--- playground/chakra-ui-playground/src/app/shared/links.ts
@@ -1,3 +1,0 @@
-import { defineLinks } from "rwsdk/router";
-
-export const link = defineLinks(["/"]);
