--- playground/chakra-ui-playground/src/app/styles.css
@@ -1,10 +1,0 @@
-body {
-  min-height: 100vh;
-  padding: 0;
-}
-
-#root,
-#hydrate-root {
-  width: 100vw;
-  height: 100vh;
-}
