--- playground/chakra-ui-playground/src/client.tsx
@@ -1,3 +1,0 @@
-import { initClient } from "rwsdk/client";
-
-initClient();
