--- playground/chakra-ui-playground/src/worker.tsx
@@ -1,17 +1,0 @@
-import { defineApp } from "rwsdk/worker";
-import { render, route } from "rwsdk/router";
-
-import { Document } from "@/app/Document";
-import { Home } from "@/app/pages/Home";
-import { setCommonHeaders } from "@/app/headers";
-
-export type AppContext = {};
-
-export default defineApp([
-  setCommonHeaders(),
-  ({ ctx }) => {
-    // setup ctx here
-    ctx;
-  },
-  render(Document, [route("/", Home)]),
-]);
