--- playground/chakra-ui-playground/types/rw.d.ts
@@ -1,5 +1,0 @@
-import { AppContext } from "../src/worker";
-
-declare module "rwsdk/worker" {
-  interface DefaultAppContext extends AppContext {}
-}
