--- playground/chakra-ui-playground/types/vite.d.ts
@@ -1,4 +1,0 @@
-declare module "*?url" {
-  const result: string;
-  export default result;
-}
