--- playground/chakra-ui-playground/vite.config.mts
@@ -1,12 +1,0 @@
-import { defineConfig } from "vite";
-import { redwood } from "rwsdk/vite";
-import { cloudflare } from "@cloudflare/vite-plugin";
-
-export default defineConfig({
-  plugins: [
-    cloudflare({
-      viteEnvironment: { name: "worker" },
-    }),
-    redwood(),
-  ],
-});
