--- playground/chakra-ui/.gitignore
@@ -1,67 +1,0 @@
-# Node modules
-node_modules
-
-# Logs
-logs
-*.log
-npm-debug.log*
-pnpm-debug.log*
-
-# Environment variables
-.env
-.dev.vars
-
-# Vite build output
-dist
-
-# TypeScript
-*.tsbuildinfo
-
-# IDEs and editors
-.vscode/
-.idea/
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw?
-
-# MacOS
-.DS_Store
-
-# Optional npm cache directory
-.npm
-
-# Optional eslint cache
-.eslintcache
-
-# Optional stylelint cache
-.stylelintcache
-
-# Optional REPL history
-.node_repl_history
-
-# Output of 'npm pack'
-*.tgz
-
-# pnpm store directory
-.pnpm-store
-
-# dotenv environment variables file
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-# Vite cache
-.vite
-
-# Coverage directory used by tools like istanbul
-coverage
-
-# Temporary files
-*.tmp
-*.temp
-
-# Wrangler
-.wrangler
