+++ playground/chakra-ui/src/app/Document.tsx
@@ -6,7 +6,7 @@ export const Document: React.FC<{ children: React.ReactNode }> = ({
     <head>
       <meta charSet="utf-8" />
       <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <title>@redwoodjs/starter-minimal</title>
+      <title>Chakra UI Playground - RedwoodSDK</title>
       <link rel="modulepreload" href="/src/client.tsx" />
       <link rel="stylesheet" href={stylesUrl} />
       <link
